refactor(feishu): share block markdown conversion
